Updated to create the content pages
If a content01.html - content09.html page is requested then it is
returned if it does exist.
If it does not exist, then the page is generated and written, and a
slightly modified content is returned to the user that created the page
letting them know that the page they are looking at was generated.

Subsequent visits to the page will return the contentxx.html page which
does not say that it was generated